requirements.txt
boto==2.35.1
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.2
django-import-export==0.2.6
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==19.1.1
Pillow==2.7.0
psycopg2==2.5.4
static3==0.5.1
tablib==0.10.0
